extract edit logic into dedicated handler
